KEYCLOAK-835 Move AdapterConstants to different package
