+++ core/src/main/java/org/keycloak/constants/GenericConstants.java
@@ -1,4 +1,4 @@
-package org.keycloak.util;
+package org.keycloak.constants;
 
 /**
  * @author <a href="mailto:mposolda@redhat.com">Marek Posolda</a>
