+++ core/src/main/java/org/keycloak/constants/ServiceUrlConstants.java
@@ -1,4 +1,4 @@
-package org.keycloak;
+package org.keycloak.constants;
 
 /**
  * @author <a href="mailto:bill@burkecentral.com">Bill Burke</a>
